requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backtrader==1.9.78.123
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
dateparser==1.1.8
fonttools==4.39.4
frozenlist==1.3.3
hyperlink==21.0.0
idna==3.4
importlib-resources==5.12.0
incremental==22.10.0
kiwisolver==1.4.4
matplotlib==3.7.1
multidict==6.0.4
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
PyAlgoTrade==0.20
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.17
pyOpenSSL==23.1.1
pyparsing==3.0.9
python-binance==1.0.17
python-dateutil==2.8.2
pytz==2023.3
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
retrying==1.3.4
scipy==1.10.1
service-identity==21.1.0
six==1.16.0
tornado==6.3.2
tweepy==4.14.0
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
ujson==5.7.0
urllib3==2.0.2
websockets==11.0.3
ws4py==0.5.1
yarl==1.9.2
zipp==3.15.0
zope.interface==6.0
